Added new network run mode
Whitebox now has the potential to use a separate directory for storing
resources (including user settings) and log files when the user does
not have write permission on the application directory. This allows for
network installation of the program and multiple-users. I also added
menu functions for viewing and deleting the log files.